feat: move all doors into new item rendering api
